refactor: move modal to new component
